refactor: change file names in the configuration dir
