+++ GANDLF/Configuration/model_config.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, model_validator, Field, AliasChoices, ConfigDict
 from typing_extensions import Self, Literal, Optional
 from typing import Union
-from GANDLF.Configuration.Parameters.validators import (
+from GANDLF.Configuration.validators import (
     validate_class_list,
     validate_norm_type,
 )
